Processing sketch to save a .svg file
This Processing sketch uses a single color for your custom hidden message pattern. See comments in sketch to customize the message. Note: All .svg glyph files in the data folder must be downloaded and saved prior to running the sketch (.pde and all .svg files should be saved in the same folder). This is used to generate an svg file that can be printed for hand embroidery of any size or converted into the necessary file type for a your embroidery machine.