Processing sketch to save a .png reference image
This Processing sketch allows for randomized color array or single color for your custom hidden message. See comments in sketch to customize. Press 's' on your keyboard to save as a .png directly to your desktop. Note: All .svg glyph files in the data folder must be downloaded and saved prior to running the sketch (.pde and all .svg files should be saved in the same folder).