Processing sketch to save a .png file
This Processing sketch uses a single color or randomized multi color array for your custom hidden message pattern. See comments in sketch to customize the message. Note: All .svg glyph files in the data folder must be downloaded and saved prior to running the sketch (.pde and all .svg files should be saved in the same folder). This is used to generate a png file that can be used as a reference image for your hand embroidery. Press 's' on your keyboard to save the image after running the script.